Try out Plug.Router
The previous solution provided us with an answer to a page in a very raw
way.  The intention is for the Router to dispatch the request to another
service.  A DSL of some sort is to be preferred.  With this commit,
we're trying out Plug.Router.  This is a first start on getting these
things up and running.