Extract Proxy from Dispatcher
The Proxy module now takes care of shipping the request/response, the
Dispatcher makes the connection and sends the correct request to the
correct location.  The current contents are still mocked, but it should
be fairly clear what an implementation of the Dispatcher will look
like.

If we assume Dockers in a future release, then we can start implementing
the routes by base path and hostname.  With a minimal abstraction, we
should then be able to tackle the dispatching in a declarative way.