Don't stop the server after starting
Without the timer, running of start would exit immediately.  This would
also mean that the application is 'finished' and stops accepting
connections.  This seems to work when running elixir interactively, as
you have to start everything manually, but it doesn't operate the way
you want it to when running it from elixir-server.