Add dockerfile
With the dockerfile, a standard version of the dispatcher can be built.
We intend to update this configuration in the near future, so this also
contains the docker-compose.  With such a construction, the dispatcher
can be used as the place to configure what is running where, and to
boot it all up.