Add basic example docker-compose file
docker-compose is the way in which we boot everything up by default.
It's a simple way for ensuring all necessary dockers get downloaded and
installed.  In the current way of deployment, it will be altered by the
end-user in order to set everything up.  It was already documented in
the README, however it should receive more content so the stub is
clearer for the end-users.