Support for matching on hostname
The host key now allows matching on hostname.  The host is extracted
into keyed segments so they can be matched later on.  The general idea
is the same as the idea of the URL matching, though currently only
supported by an array.  The first element of the array is the
last (and most broad) element of the domain name.

future.redpencil.io would thus become ["io", "redpencil", "future"].