Module attribute manipulation and drop last_match requirement
This constitutes a rewrite of the macro capturing system.  The
attribute manipulation seemingly required us to manipulate things just
before compilation.  I currently doubt we determined that to be the
actual case.  It seems likely, but not guaranteed.  This new
implementation has the upside that we don't require last_match to be
entered.  As I suspect this to be something which may be forgotten
rather often, leading to errors in a late path, it's nice we can drop
this.