Adding documentation for matching on host
Matching on host has been added to the documentation.

Whilst writing the docs, we've been wondering if using the same key
for the string variant as well as for the array variant is a good
idea.  It seems sane to go with it, but we should question the
implementation before giving it a final go.